Add a more generalized uw rdt capture widget
