+++ rdt-open-reader/src/main/java/io/ona/rdt/interactor/RDTJsonFormInteractor.java
@@ -4,7 +4,8 @@ import com.vijay.jsonwizard.interactors.JsonFormInteractor;
 
 import org.json.JSONObject;
 
-import io.ona.rdt.widget.CustomRDTCaptureFactory;
+import io.ona.rdt.widget.UWCovidRDTCaptureFactory;
+import io.ona.rdt.widget.UWMalariaRDTCaptureFactory;
 import io.ona.rdt.widget.RDTBarcodeFactory;
 import io.ona.rdt.widget.RDTCountdownTimerFactory;
 import io.ona.rdt.widget.RDTExpirationDateReaderFactory;
@@ -16,6 +17,7 @@ import static com.vijay.jsonwizard.constants.JsonFormConstants.COUNTDOWN_TIMER;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.GPS;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.LABEL;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.RDT_CAPTURE;
+import static io.ona.rdt.util.Constants.Widget.UW_COVID_RDT_CAPTURE;
 import static io.ona.rdt.widget.RDTExpirationDateReaderFactory.EXPIRATION_DATE_CAPTURE;
 
 /**
@@ -37,9 +39,10 @@ public class RDTJsonFormInteractor extends JsonFormInteractor {
         map.put(BARCODE, new RDTBarcodeFactory());
         map.put(LABEL, new RDTLabelFactory());
         map.put(EXPIRATION_DATE_CAPTURE, new RDTExpirationDateReaderFactory());
-        map.put(RDT_CAPTURE, new CustomRDTCaptureFactory());
+        map.put(RDT_CAPTURE, new UWMalariaRDTCaptureFactory());
         map.put(COUNTDOWN_TIMER, new RDTCountdownTimerFactory());
         map.put(GPS, new RDTGpsFactory());
+        map.put(UW_COVID_RDT_CAPTURE, new UWCovidRDTCaptureFactory());
     }
 
     public void saveForm(JSONObject jsonForm) {
